+++ mjx/mujoco/mjx/_src/types.py
@@ -763,11 +763,13 @@ class Model(PyTreeNode):
     tendon_adr: address of first object in tendon's path      (ntendon,)
     tendon_num: number of objects in tendon's path            (ntendon,)
     tendon_limited: does tendon have length limits            (ntendon,)
+    tendon_actfrclimited: tendon has actuator force limits    (ntendon,)
     tendon_solref_lim: constraint solver reference: limit     (ntendon, mjNREF)
     tendon_solimp_lim: constraint solver impedance: limit     (ntendon, mjNIMP)
     tendon_solref_fri: constraint solver reference: friction  (ntendon, mjNREF)
     tendon_solimp_fri: constraint solver impedance: friction  (ntendon, mjNIMP)
     tendon_range: tendon length limits                        (ntendon, 2)
+    tendon_actfrcrange: tendon actuator force limits          (ntendon, 2)
     tendon_margin: min distance for limit detection           (ntendon,)
     tendon_stiffness: stiffness coefficient                   (ntendon,)
     tendon_damping: damping coefficient                       (ntendon,)
@@ -1106,11 +1108,13 @@ class Model(PyTreeNode):
   tendon_adr: np.ndarray
   tendon_num: np.ndarray
   tendon_limited: np.ndarray
+  tendon_actfrclimited: np.ndarray
   tendon_solref_lim: jax.Array
   tendon_solimp_lim: jax.Array
   tendon_solref_fri: jax.Array
   tendon_solimp_fri: jax.Array
   tendon_range: jax.Array
+  tendon_actfrcrange: jax.Array
   tendon_margin: jax.Array
   tendon_stiffness: jax.Array
   tendon_damping: jax.Array